% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_densities.R
\name{logpi_lm}
\alias{logpi_lm}
\title{Log Posterior Density for Linear Regression}
\usage{
logpi_lm(beta = c(0, 0), prior = 10^(-3) * numeric(length(b)))
}
\arguments{
\item{prior}{}
}
\value{

}
\description{
Evaluates the log posterior density for linear regression using a particular prior.
Code is modified from course materials from the Boston University course MA578 as taught by Yves Atchade.
}
\examples{

x <- matrix(rnorm(100))
x <- cbind(1,x)
y <- x\%*\%c(5,1)+rnorm(50,0,1/2)
logpi_logistic(c(5,1),y,x)


}
