% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_densities.R
\name{logpi_logistic}
\alias{logpi_logistic}
\title{Log Posterior Density for Logistic Regression}
\usage{
logpi_logistic(beta, y, X, c_prior = 100)
}
\arguments{
\item{beta}{Beta values}

\item{y}{Binary vector on which to fit logistic regression}

\item{X}{Matrix of continous predictors on which to fit logistic regression}

\item{c_prior}{Paramater for prior.}
}
\value{

}
\description{
Evaluates the log posterior density for logistic regression using a particular prior.
Code is modified from course materials from the Boston University course MA578 as taught by Yves Atchade.
}
\examples{

x <- matrix(rnorm(100))
y <- rbinom(100,1,exp(x)/(1+exp(x)))
logpi_logistic(0,y,x)

}
