% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling_functions.R
\name{sample_maximal_coupling}
\alias{sample_maximal_coupling}
\title{Sample Maximal Coupling}
\usage{
sample_maximal_coupling(psample, qsample, pdensity, qdensity)
}
\arguments{
\item{psample}{A function which produces 1 sample from the distribution p}

\item{qsample}{A function 1 sample from the distribution q}

\item{pdensity}{A function of double x returning the log density of x from distribution p}

\item{qdensity}{A function of double x returning the log density of x from distribution q}
}
\value{

}
\description{
Uses maximal coupling to produce a single sample from distriubtions p and q.
}
