% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvn_functions.R
\name{sample_mvn}
\alias{sample_mvn}
\title{Multivariate Normal Sample}
\usage{
sample_mvn(mu, sigma = diag(length(mu)), R = t(chol(sigma)), log_scaling = 0)
}
\arguments{
\item{mu}{Mean of distribution to sample from}

\item{sigma}{Covariance matrix \eqn{\Sigma} to sample from}

\item{R}{Cholesky factor of \eqn{\Sigma}}

\item{log_scaling}{Paramater \eqn{\sigma} in \eqn{e^{2\sigma}\Sigma}}
}
\value{

}
\description{
Function to generate a single sample from the multivariate normal distribution.
Allows easy parameterization of the covariance matrix as \eqn{e^{2\sigma}\Sigma}.
}
