% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupling_functions.R
\name{sample_reflection_maximal}
\alias{sample_reflection_maximal}
\title{Sample using Reflection Maximal Sampling}
\usage{
sample_reflection_maximal(
  mu_p,
  mu_q,
  sigma,
  s_sample = function() {     sample_mvn(numeric(length(mu_p))) },
  s_density = function(x) {     log_density_mvn(x) }
)
}
\arguments{
\item{mu_p}{Mean of the distribution p.}

\item{mu_q}{Mean of the distribution q.}

\item{s_sample}{A function which returns one sample from the distribution s. Defaults to standard multivariate normal.}

\item{s_density}{A function of x returning the log density of x from the distribution s. Defaults to standard multivariate normal.}

\item{Sigma}{Covariance matrix of distributions p and q.}
}
\value{

}
\description{
See Unbiased Markov chain Monte Carlo with couplings for details.
}
\examples{

mu_p = -1
mu_q = 1
sigma = diag(1)





niter <- 1000
rmat <- matrix(nrow = niter, ncol = 2)

for(i in 1:niter){
r <- sample_reflection_maximal(mu_p,mu_q,sigma)
rmat[i,1] <- r$x
rmat[i,2] <- r$y
}

hist(c(rmat[,1],rmat[,2]))
plot(rmat[,1]~rmat[,2])



}
