% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_averaged_estimator.R
\name{time_averaged_estimator}
\alias{time_averaged_estimator}
\title{Unbiased Time averaged estimator}
\usage{
time_averaged_estimator(res, m, k, tau, h = function(x) {
    x
})
}
\arguments{
\item{res}{A matrix returned by coupled_ARWM().}

\item{tau}{Tau as returned by coupled_ARWM(). Note that this is the first INDEX that the chain meets, so the meeting time is tau-1.}

\item{h}{A vectorized function.}
}
\value{

}
\description{
Unbiased Time averaged estimator
}
