% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupled_ARWM.R
\name{coupled_ARWM}
\alias{coupled_ARWM}
\title{Coupled Adaptive Random Walk Metropolis Algorithm}
\usage{
coupled_ARWM(
  Niter,
  y,
  X,
  logpi = logpi_logistic,
  type = c("maximal", "reflection maximal", "none"),
  acceptance_probability = 0.3,
  beta_init = numeric(length(X[1, ])),
  beta_couple_init = numeric(length(X[1, ])),
  C_init = diag(length(X[1, ])),
  lsig_init = 1,
  mu_init = numeric(length(X[1, ]))
)
}
\arguments{
\item{Niter}{Number of iterations for the algorithm to run}

\item{X}{}
}
\value{

}
\description{
Given a log posterior density, generates samples using coupled adaptive random walk metropolis.
Code is modified from course materials from the Boston University course MA578 as taught by Yves Atchade.
}
\examples{

x <- matrix(rnorm(100))
y <- rbinom(100,1,exp(x)/(1+exp(x)))



r <- coupled_ARWM(1000,y,x, type = "reflection maximal", acceptance_probability = 0.3)

n <- 1000
t <- numeric(n)
for(i in 1:n){
r <- coupled_ARWM(1,y,x, type = "reflection maximal")
t[i] <- r$tau

if(i \%\% 10 == 0){print(i)}

}



}
